Make path::resolve a method on ExtCtxt
